Added `assessCreditLine` function and refactors.
- Assesment function works by getting called manually, and looking at
a particular credit line. It uses the nextDueBlock, and does all
appropriate checking to pull prepayment balance and use it to pay off accrued
balances.
- Also refactored so that both `pay` and `assessCreditLine` use a single
"handlePayment" function which deals with the logic for updating accounting
vars on the CreditLine and sending the correct amounts to the pool.
- The only nit there is there's an if statement for whether we want to the payment
to be able to count towards "extra" principal. In the "pay" case we do, but in
the assessment case we do not (at least currently we don't)
- Also refactored to use SafeMath everywhere. Realized we should totally be doing that,
because we do some subtractions that theoretically could go negative (really shouldn't, but who knows),
and we'd want those to error, rather than set people's balances to like a bajillion.