Add in a threshold for the assets/liabilities check.
- We now allow for up to a $1 threshold of difference between the
pool assets and liabilities.
- The reason is that we can ocassionally, depending on the share price,
end up with super small discrepancies that are strictly due to rounding,
and nothing actually wrong with the protocol (false positives).
- The underlying reason for this is that Fidu and USDC have differing decimal
precision (18, and 6, respectively), so getting them into the same units can cause
discrepancies.
- Also, the implementation here is slightly less than ideal, due to lack of time.
Really this calculation could or should be put into a separate library or something.
- But for sake of time (we have the audit actively under way), I'm doing a small bit of duplication.