[Feature] Factory Creation of Tranched Pools (#244)
* Got a basic version of the create2 working.

- Our CreditLineFactory now creates Pools as well.
- It does this via create2. We then use the create2 magic to
determine if the caller of the mint function in PoolTokens is a validly
created Pool.
- We don't have a working implementation, so I'm using a stand-in of the Borrower contract.
- This also still hasn't solved what we're using for the salt. or how that salt
will be known to the PoolTokens contract, but we'll figure that out next.

* Hooked up the TranchedPool with the PoolTokens.

- Now, we are actually using the TranchedPool when we create a Pool
- In so doing, a number of other changes got made, including adding
creditLine, and createdAt to the ITranchedPool, as well as adding
an IV2CreditLine which can be used by the TranchedPool.
- Also we're now using a workable salt when creating the Pools.
Specifically, we are using the borrower address + createdAt (as block.timestamp)

* Got minimal proxy deploys working.

- Instead of deploying the full TranchedPool every time, we now
deploy minimal proxies that point to a single implementation.
- This will save us quite a bit of gas costs as we deploy many of these.
- Also broke out some helpers into a DeployHelpers library, for things like
create2Deploy, and calculating the create2 address.
- Also got more tests passing, and added a way to decode raw logs, which is
necessary for using logs that are emitted from a different contract than
the one you're working with directly. (ie. PoolTokens emits it when you called the Pool)

* A few cleanups from self review.

* Fix tests

* Move validPool to the factory

* Added a TODO for the min proxy bug.

Co-authored-by: Sanjay Raveendran <sanjayprabhu@gmail.com>