[Fix] Asset and liability mismatch when transferring USDC directly to Pool (#242)
Fixes an issue where doing an ERC20 transfer into Pool would break any subsequent
Pool deposits and withdrawals. This was due to a check that assets == liabilities. We
relax this check, requiring only that assets >= liabilities i.e. that FIDU is *at
least* backed by enough USDC either present in the Pool or held in loans.