[chore] Static geolocation banners (#455)
Hardcode banners to indicate investments are for non-US only. Also, fix styles so
they match master.

The first open pool will be non-US only. By hardcoding, we defer dealing with edge cases
like ensuring that a user KYCed in the US is not seeing the non-US banner or
supporting pools that want to only allow US or non-US