[fix] Senior pool APY (#510)
Migrated tranched pools do not have InvestmentMadeInJunior events, as the position tokens
were minted directly. So they were not accounted for in getEestimatedTotalInterest.
Manually include these in the pools that should be considered for estimated interest
using the metadata store. This works because the senior pool was guaranteed by the
migration contract to have invested as junior.