[fix] V2 pool launch fixes (#513)
* [fix] Optimize splits

Don't consider splits for tokens with zero value. This makes gas costs lower

* [fix] Backer limits

Consider principalRedeemed when determining backer deposit amount. Otherwise, we
don't properly account for withdrawals.

* [fix] UI flickers after tranched pool action

The useAsync pattern is leading to re-renders where tranchedPool or backer is
undefined (when the result is still loading). Work around this by using a separate
state that only gets set when the result is loaded.