Support deploying MerkleDistributor contract using Merkle root calculated for specified recipients, on murmuration.
This approach, of using a predefined set of accounts to use as recipients
of grants claimable via the MerkleDistributor contract, was a more expedient
solution than trying to use our new "dev tools" pattern to allow the user of
murmuration to bring about the same outcome dynamically. That's because
the dynamic "dev tools" approach would require somehow refreshing / updating
the `all_dev.json` file generated by hardhat-deploy that contains the ABI
of the deployed contracts. Specifically, we'd need that `all_dev.json` to
be updated to include whatever new MerkleDistributor contract was deployed
by the dev tools script. I did not find a straightforward way to do that,
as `all_dev.json` gets generated upon starting the Hardhat node.