[fix] Allow partial vesting periods (#246)
* [fix] Allow partial vesting periods

A small percentage of grants were added with vesting periods that don't evenly divide
the vesting length. Allow for this scenario, letting accounts be fully vested in the
last whole vesting period.

* Add migration script + mainnet forking test

* [fix] Set default for vestingInterval of 0

There is a single grant with a vestingInterval of 0, which would result in a division
by 0 upon claim. Set a default value of vestingLength, so they can claim at the end
of vesting.

* Fix forking test

Co-authored-by: gregan <greg@goldfinch.finance>