Set credit line's `termEndTime` only once on the first drawdown. (#695)
### What happened?
The `termEndTime` is extended by a full term length whenever there is a drawdown on a 0 balance credit line. We change the drawdown function so the `termEndTime` is only extended once on the initial drawdown. Once initialized, it should not change.

### Why?
Without this restriction a borrower can arbitrarily increase the termEndTime by repeatedly drawing down and paying back the full balance. This would indefinitely delay the time at which the principal is owed.

### How has this been tested?
Unit tests

**Checklist**
- [x] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)