[chore] Remove outdated protocol_config.json (#721)
* [chore] Remove outdated protocol_config.json

We originally used protocol_config.json as the canonical values to configure the
protocol. But it ended up being cumbersome in practice, so we ended up seeding new
GoldfinchConfig deployments with the older config's values, making the json file no
longer canonical. Now that we have a singleton, upgradeable GoldfinchConfig, its
state variables are the canonical config. We may want to revisit what we consider the
canonical config values, but it made sense to remove the json file to avoid confusion
in the meantime.

* Remove Deployments test

* Remove unused npm scripts