[chore] Remove getContract in favor of explicit get(Truffle|Ethers)Contract (#724)
The getContract function was written first and required passing both ethers and truffle
types as well as contract provider type, in order to get back a typed contract for
one of those frameworks. This made usage a bit verbose. Remove in favor of the explicit functions
that doesn't require callers to have knowledge of both frameworks / types.