Implement script for burning UIDs (#431)
### What happened?
Added a script to burn UIDs and optionally allow the user to re-mint with a new address. Here is a summary of the involved entities and the data flows between them:

**burnUID**
The main entry-point of the script. Invoked from the command line for prod or POSTed to at `localhost:4000/burnUID` for localhost.

**Relayer**
I created a [relayer](https://defender.openzeppelin.com/#/relay/46bdacb4-bc7e-475c-a3f1-b0ab77b98873/settings) on defender called Unique Identity Burner. It serves as a trusted signer that is allowed to call the on-chain burn function and also call the _destroyUser_ cloud function

**destroyUser cloud function**
A new cloud function that deletes from the _users_ collection and writes to the _destroyedUsers_ collection. Only allowed signers can call this function. The verification is done with a new auth mechanism called "signatureWithAllowList"

**user and destroyedUsers stores**
Firebase stores. The _destroyedUsers_ store is new and acts as the "soft delete". We delete from the user store but write the deleted data to the destroyedUsers store for record keeping purposes

**UniqueIdentity**
On chain UID contract. We call burn on this contract

**Persona**
We call the persona API to update the account's reference ID (wallet address) if the `prepareRemintAccount` variable is present.

![image](https://user-images.githubusercontent.com/19829228/171978033-a8ad3f83-72d6-4057-9147-2260a25e43a7.png)

### Why?
It is sometimes useful to burn a user's UID so they can re-mint a UID of a different type. 

### How has this been tested?
- [x] burnUID first time, no remint account
  - [x] UID burned on chain
  - [x] row deleted from users store
  - [x] row created in destroyedUsers store
  - [x] I can re-kyc and remint with same address
- [x] burnUID not first time
  - [x] UID burned on chain
  - [x] row deleted from users store
  - [x] append to existing row in destroyedUsers store
  - [x] I can re-kyc and remint with same address
- [x] burnUID with remint account
  - [x] UID burned on chain
  - [x] row deleted from users store
  - [x] row created in destroyedUsers store
  - [x] I can re-kyc and remint with remint address

**Checklist**
- [ ] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)