remove go changes (#1198)
### What happened?
Remove the Go changes that allow a user to ERC1155 approve a contract to act with their UID

### Why?
I added this in because I thought the [august-audit PR ](https://github.com/warbler-labs/mono/pull/800)with the same Go changes would be merged by now but it isn't. We should remove it to reduce the scope of the audit.

### How has this been tested?
Existing unit tests pass. The unit tests for Go have been commented out with a TODO to uncomment them when august-audit is merged.

**Checklist**
- [ ] I have reviewed this code myself
- [ ] Screenshots of the key change are present (if this is a front-end PR)